gateway/requirements.txt
fastapi
httpx
uvicorn[standard]
pyjwt
prometheus-fastapi-instrumentator